Making ModuleWrapper abstract to allow multiple implementations
